removed unnecessary input parameters, refactored inout into two files
