--- examples/atomic.yml
@@ -1,49 +1,0 @@
----
-control:
-  sampling:         rejection       # only rejection sampling is implemented currently
-  xsection:         Full
-  beam:
-    sample_electrons:  5e6         # total number of macroelectrons electrons to sample
-    sample_batch_size: 1e6         # number of macroelectrons per batch
-
-  # rescaling of pulse duration
-  laser:
-    pulse_rescale_bias:    1.      # > 1 for better sampling at the cost of larger natural bandwidth
-  detector:
-    a0_freq_correction: False       # include a0-dependent frequency red-shift 
-
-
-unit:
-  momentum:    keV
-  position:    micron
-
-
-beam:
-  gamma:        1.0
-  energyspread: 0.0
-  emittanceX:   0.0           # mm mrad
-  emittanceY:   0.0           # mm mrad
-  sigmaX:       0.1           # micron, transverse beam size
-  sigmaY:       0.1           # micron, transverse beam size
-  sigmaL:       0.1           # micron, longitduinal beam size
-  beam_charge:  100e-12       # beam charge in Coulomb
-  beam_focus_z: 0             # Focus position of electron beam, 0 = ICS interaction point, in microns
-  baseline:     1.0           # Distance between Compton Scattering and detector in micron
-
-
-laser:
-  a0:           0.1           # normalized laser vector potential
-  omega0:       3.e5          # eV, laser frequency 
-  TFWHM:        1.0e-4        # fs, FWHM laser pulse duration
-  pulse:        cos2          # pulse envelope function, only cos2 supported so far
-  w0:           100.          # micron, rms laser spot size
-  polangle:     0.            # laser polarization angle from x-axis in rad
-  poldegree:    1.            # laser polarization degree
-
-
-detector:                           # spectrum calculation
-  pdim:  3                          # number of detector dimensions for spectrum calculation
-  omega: [0.5e5,3.3e5]              # [omegamin,omegamax] , eV
-  theta: [0.0,3.14159265]           # [thetamin,thetamax] , rad
-  phi:   [0.0,6.28318530]           # [phimin,phimax]     , rad
-
